⚡️ setup: パスを絶対参照に変更, realpathをPWDに変更

+++ setup.sh
@@ -2,25 +2,25 @@
 
 # シンボリックリンクを作成するファイルとパスのリスト
 files_and_paths=(
-  ".bashrc:~/.bashrc"
+  ".bashrc:$HOME/.bashrc"
 )
 
 # シンボリックリンクを作成する関数
 create_symlink() {
-  local source_file=$(realpath $1)
+  local source_file="$PWD/$1"
   local destination_path=$2
 
   # 退避先のファイル名
   backup_file="${destination_path}.bak"
 
   # バックアップファイルが存在する場合は削除
-  if [ -e "$destination_path" ]; then
+  if [ -e "$destination_path" ] && [ ! -L "$destination_path" ]; then
     # バックアップ
     mv "$destination_path" "$backup_file"
   fi
 
-  # シンボリックリンクの作成
-  ln -s "$source_file" "$destination_path"
+  # シンボリックリンクの作成（既存のリンクは強制上書き）
+  ln -sf "$source_file" "$destination_path"
 }
 
 # ファイルとパスのリストをループしてシンボリックリンクを作成
